web_2025/requirements.txt
Django==5.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0